Compilation fixes

+++ Renderer/Renderer/RayCasting/Geometry/RayTraceScene.cpp
@@ -323,12 +323,12 @@ void C_RayTraceScene::BuildScene()
 				if (mat.textureIndex != -1)
 				{
 					m_Textures.emplace_back(rm.LoadResource<TextureResource>(std::filesystem::path(meshHandle.GetResource().GetTextureNames()[mat.textureIndex])));
-					mat.textureIndex = m_Textures.size() - 1;
+					mat.textureIndex = static_cast<int>(m_Textures.size() - 1);
 				}
 				if (mat.noramlTextureIndex != -1)
 				{
 					m_Textures.emplace_back(rm.LoadResource<TextureResource>(std::filesystem::path(meshHandle.GetResource().GetTextureNames()[mat.textureIndex])));
-					mat.noramlTextureIndex = m_Textures.size() - 1;
+					mat.noramlTextureIndex = static_cast<int>(m_Textures.size() - 1);
 				}
 				mat.shininess		   = 0.f;
 				AddMesh(mesh, mat);
